+++ main.py
@@ -1,331 +1,8 @@
-import tkinter as tk
-from tkinter import *
-from tkinter import ttk
-from tkinter_custom_button import TkinterCustomButton
-
-
-class MainWindow:
-
-    def __init__(self):
-        global root
-
-        for widget in root.winfo_children()[1:]:
-            widget.destroy()
-
-        #root.geometry("600x600")
-        #root.minsize(750, 600)
-        root.configure(background="#2a3439")
-
-class ResultsPage:
-
-    def __init__(self):
-        global root
-
-        for widget in root.winfo_children()[1:]:
-            widget.destroy()
-
-        # root.geometry("600x600")
-        # root.minsize(750, 600)
-        root.configure(background="#2a3439")
-
-        # <editor-fold desc="Results GUI">
-        # Frame for whole results page
-        results_frame = tk.Frame(root, bg="#2a3439")
-        results_frame.place(relx=0.5, rely=0.1, anchor="n")
-        results_frame.config(height=root.winfo_height(), width=root.winfo_width())
-
-        # Container for results
-        results_container = tk.Frame(results_frame, bg="#1F262A", borderwidth = 2)
-        results_container.place(relx=0.5, rely=0.1, anchor="n")
-        results_container.config(relief=RIDGE)
-
-        # Result examples
-        results_example1 = tk.Frame(results_container, bg="#2a3439")
-        results_example1.place(relx=0.5, rely=0.02, anchor="n")
-        results_example1.config(height=50, width=900)
-        results_example1_label = tk.Label(results_example1, text='Software 1',  font=14, bg="#2a3439", fg="#5B676D")
-        results_example1_label.place(relx=0.01, rely=0.5, anchor="w")
-
-        results_example2 = tk.Frame(results_container, bg="#2a3439")
-        results_example2.place(relx=0.5, rely=0.02, anchor="n")
-        results_example2.config(height=50, width=900)
-        results_example2_label = tk.Label(results_example2, text='Software 2', font=14, bg="#2a3439", fg="#5B676D")
-        results_example2_label.place(relx=0.01, rely=0.5, anchor="w")
-
-        results_example3 = tk.Frame(results_container, bg="#2a3439")
-        results_example3.place(relx=0.5, rely=0.02, anchor="n")
-        results_example3.config(height=50, width=900)
-        results_example3_label = tk.Label(results_example3, text='Software 3', font=14, bg="#2a3439", fg="#5B676D")
-        results_example3_label.place(relx=0.01, rely=0.5, anchor="w")
-
-        results_example4 = tk.Frame(results_container, bg="#2a3439")
-        results_example4.place(relx=0.5, rely=0.02, anchor="n")
-        results_example4.config(height=50, width=900)
-        results_example4_label = tk.Label(results_example4, text='Software 4', font=14, bg="#2a3439", fg="#5B676D")
-        results_example4_label.place(relx=0.01, rely=0.5, anchor="w")
-
-        results_example5 = tk.Frame(results_container, bg="#2a3439")
-        results_example5.place(relx=0.5, rely=0.02, anchor="n")
-        results_example5.config(height=50, width=900)
-        results_example5_label = tk.Label(results_example5, text='Software 5', font=14, bg="#2a3439", fg="#5B676D")
-        results_example5_label.place(relx=0.01, rely=0.5, anchor="w")
-
-        results_example6 = tk.Frame(results_container, bg="#2a3439")
-        results_example6.place(relx=0.5, rely=0.02, anchor="n")
-        results_example6.config(height=50, width=900)
-        results_example6_label = tk.Label(results_example6, text='Software 6', font=14, bg="#2a3439", fg="#5B676D")
-        results_example6_label.place(relx=0.01, rely=0.5, anchor="w")
-
-        # Align results in a grid
-        results_example1.grid(row=0, column=0, padx=10, pady=5)
-        results_example2.grid(row=1, column=0, padx=10, pady=5)
-        results_example3.grid(row=2, column=0, padx=10, pady=5)
-        results_example4.grid(row=3, column=0, padx=10, pady=5)
-        results_example5.grid(row=4, column=0, padx=10, pady=5)
-        results_example6.grid(row=5, column=0, padx=10, pady=5)
-        # </editor-fold>
-
-        # <editor-fold desc="Results Buttons">
-        update_all_button = TkinterCustomButton(master=results_frame,
-                                          fg_color="#848689",
-                                          hover_color="#1F262A",
-                                          text_font="Bold, 14",
-                                          text="Update All",
-                                          text_color="white",
-                                          corner_radius=10,
-                                          width=200,
-                                          height=75,
-                                          hover=True,
-                                          command=lambda: None)
-        update_all_button.place(relx=0.25, rely=0.8, anchor="center")
-
-        update_selected_button = TkinterCustomButton(master=results_frame,
-                                                fg_color="#8797AF",
-                                                hover_color="#1F262A",
-                                                text_font="Bold, 14",
-                                                text="Update Selected",
-                                                text_color="white",
-                                                corner_radius=10,
-                                                width=200,
-                                                height=75,
-                                                hover=True,
-                                                command=lambda: None)
-        update_selected_button.place(relx=0.5, rely=0.8, anchor="center")
-
-        cancel_button = TkinterCustomButton(master=results_frame,
-                                                fg_color="#5F4866",
-                                                hover_color="#1F262A",
-                                                text_font="Bold, 14",
-                                                text="Cancel",
-                                                text_color="white",
-                                                corner_radius=10,
-                                                width=100,
-                                                height=50,
-                                                hover=True,
-                                                command=lambda: None)
-        cancel_button.place(relx=0.70, rely=0.8, anchor="center")
-        # </editor-fold>
-
-
-class HelpPage:
-
-    def __init__(self):
-        global root
-
-        for widget in root.winfo_children()[1:]:
-            widget.destroy()
-
-        # root.geometry("600x600")
-        # root.minsize(750, 600)
-        root.configure(background="#2a3439")
-
-        # <editor-fold desc="Results GUI">
-        # Frame for whole results page
-        help_frame = tk.Frame(root, bg="#2a3439")
-        help_frame.place(relx=0.5, rely=0.1, anchor="n")
-        help_frame.config(height=root.winfo_height(), width=root.winfo_width())
-
-        # Container for results
-        help_container = tk.Frame(help_frame, bg="#1F262A", borderwidth=2)
-        help_container.place(relx=0.5, rely=0.1, anchor="n")
-        help_container.config(relief=RIDGE)
-
-        # Help tip examples
-        help_example1 = tk.Frame(help_container, bg="#2a3439")
-        help_example1.place(relx=0.5, rely=0.02, anchor="n")
-        help_example1.config(height=200, width=900)
-        help_example1_header_label = tk.Label(help_example1, text='How to use the program:', font=24, bg="#2a3439", fg="#FFFFFF")
-        help_example1_header_label.place(relx=0.01, rely=0.1, anchor="nw")
-
-        help_example2 = tk.Frame(help_container, bg="#2a3439")
-        help_example2.place(relx=0.5, rely=0.02, anchor="n")
-        help_example2.config(height=200, width=900)
-        help_example2_header_label = tk.Label(help_example2, text='How the Vulnerabilities are scored:', font=24, bg="#2a3439", fg="#FFFFFF")
-        help_example2_header_label.place(relx=0.01, rely=0.1, anchor="nw")
-
-        help_example3 = tk.Frame(help_container, bg="#2a3439")
-        help_example3.place(relx=0.5, rely=0.02, anchor="n")
-        help_example3.config(height=200, width=900)
-        help_example3_header_label = tk.Label(help_example3, text="What databases we're checking against:", font=24, bg="#2a3439", fg="#FFFFFF")
-        help_example3_header_label.place(relx=0.01, rely=0.1, anchor="nw")
-
-        # Align tips in a grid
-        help_example1.grid(row=0, column=0, padx=10, pady=5)
-        help_example2.grid(row=1, column=0, padx=10, pady=5)
-        help_example3.grid(row=2, column=0, padx=10, pady=5)
-        # </editor-fold>
-
-
-class SettingsPage:
-
-    def __init__(self):
-        global root
-
-        for widget in root.winfo_children()[1:]:
-            widget.destroy()
-
-        root.configure(background="#2a3439")
-
-        settings_frame = tk.Frame(root)
-        settings_frame.place(relx=0.5, rely=0.5, anchor='center')
-        settings_frame.config(height=300, width=500)
-        settings_frame.config(relief=RIDGE)
-
-        settings_page_label = ttk.Label(settings_frame, text='Settings Page')
-        settings_page_label.place(relx=0.5, rely=0.15, anchor='center')
-
-        set_options_frame = ttk.Frame(settings_frame)
-        set_options_frame.place(relx=0.5, rely=0.5, anchor='center')
-        set_options_frame.config(height=300, width=500)
-        set_options_frame.config(relief=RIDGE)
-        set_options_frame.config(padding=(30, 15))
-
-        set1_label = ttk.Label(set_options_frame, text='Setting 1')
-        set1_label.grid(row=0, column=0)
-        set2_label = ttk.Label(set_options_frame, text='Setting 2')
-        set2_label.grid(row=1, column=0)
-        set3_label = ttk.Label(set_options_frame, text='Setting 3')
-        set3_label.grid(row=2, column=0)
-    
-    
-class LoginPage:
-
-    def __init__(self):
-        global root
-
-        for widget in root.winfo_children()[1:]:
-            widget.destroy()
-
-        root.configure(background="#2a3439")
-
-        login_outer_frame = Frame(root, bg='#1F262A')
-        login_outer_frame.place(relx=0.5, rely=0.5, anchor='center')
-        login_outer_frame.config(height=400, width=600)
-        login_outer_frame.config(relief=RAISED)
-
-        login_inner_frame = Frame(root, background="#2a3439")
-        login_inner_frame.place(relx=0.5, rely=0.5, anchor='center')
-        login_inner_frame.config(height=300, width=500)
-
-        login_page_label = Label(root, text='Login Page', font="Bold, 20", bg='#2a3439', fg="white")
-        login_page_label.place(relx=0.5, rely=0.3, anchor='center')
-
-        username_label = Label(root, text='Username', font=15, background="#2a3439", foreground="white")
-        username_label.place(relx=0.5, rely=0.4, anchor="center")
-        username_entry = Entry(root, background="#1F262A", foreground="white", font=15)
-        username_entry.place(relx=0.5, rely=0.45, anchor='center')
-        token_label = Label(root, text='RSA Token', font=15, background="#2a3439", foreground="white")
-        token_label.place(relx=0.5, rely=0.515, anchor='center')
-        token_entry = Entry(root, background="#1F262A", foreground="white", font=15)
-        token_entry.place(relx=0.5, rely=0.565, anchor='center')
-
-        #Login Button
-        login_button = TkinterCustomButton( bg_color="#2a3439",
-                                          fg_color="#56667A",
-                                          hover_color="#AAA9AD",
-                                          text_font="Bold, 12",
-                                          text="Login",
-                                          text_color="white",
-                                          corner_radius=10,
-                                          width=80,
-                                          height=40,
-                                          hover=True,
-                                          command=lambda: MainWindow())
-        login_button.place(relx=0.45, rely=0.65, anchor='center')
-
-        #Registration Button
-        register_button = TkinterCustomButton( bg_color="#2a3439",
-                                          fg_color="#56667A",
-                                          hover_color="#AAA9AD",
-                                          text_font="Bold, 12",
-                                          text="Register",
-                                          text_color="white",
-                                          corner_radius=10,
-                                          width=80,
-                                          height=40,
-                                          hover=True,
-                                          command=lambda: RegisterPage())
-        register_button.place(relx=0.55, rely=0.65, anchor='center')
-
-        
-class RegisterPage:
-
-    def __init__(self):
-        global root
-
-        for widget in root.winfo_children()[1:]:
-            widget.destroy()
-
-        root.configure(background="#2a3439")
-
-        register_frame = Frame(root, bg='#1F262A')
-        register_frame.place(relx=0.5, rely=0.5, anchor='center')
-        register_frame.config(height=350, width=500)
-        register_frame.config(relief=RAISED)
-
-        register_title = Label(root, text="Register", background="#1F262A", foreground="white", font="Bold, 25")
-        register_title.place(relx=0.5, rely=.3, anchor='center')
-        first_name = Label(root, text="First Name", background="#1F262A", foreground="white", font=20)
-        first_name.place(relx=.34, rely=.38)
-        first_name_entry = Entry(root, background="#2a3439", foreground="white", width=25, font=20)
-        first_name_entry.place(relx=.45, rely=.38)
-        last_name = Label(root, text="Last Name", background="#1F262A", foreground="white", font=20)
-        last_name.place(relx=.34, rely=.46)
-        last_name_entry = Entry(root, background="#2a3439", foreground="white", width=25, font=20)
-        last_name_entry.place(relx=.45, rely=.46)
-        username = Label(root, text="Username", background="#1F262A", foreground="white", font=20)
-        username.place(relx=.34, rely=.54)
-        username_entry = Entry(root, background="#2a3439", foreground="white", width=25, font=20)
-        username_entry.place(relx=.45, rely=.54)
-        token = Label(root, text="RSA Token ID", background="#1F262A", foreground="white", font=20)
-        token.place(relx=.34, rely=.62)
-        token_entry = Entry(root, background="#2a3439", foreground="white", width=25, font=20)
-        token_entry.place(relx=.45, rely=.62)
-
-        #Create Account Button
-        create_button = TkinterCustomButton( bg_color="#1F262A",
-                                          fg_color="#56667A",
-                                          hover_color="#AAA9AD",
-                                          text_font= 20,
-                                          text="Create",
-                                          text_color="white",
-                                          corner_radius=10,
-                                          width=100,
-                                          height=30,
-                                          hover=True,
-                                          command=lambda: LoginPage())
-        create_button.place(relx=0.5, rely=0.73, anchor='center')
-
-
-root = tk.Tk()
+from PageClasses import *
+from PageClasses import root as root
 
 
 def create_window():
-    # root = tk.Tk()
-
-    # Initialize all windows
-    MainWindow()
-
     # Window background color
     root.configure(background="#2a3439")
 
@@ -340,82 +17,70 @@ def create_window():
 
     # remove title bar
     root.overrideredirect(True)
-
-    #
-    def move_app(e):
-        root.geometry(f'+{e.x_root}+{e.y_root}')
-
-    #
-    def quitter(e):
-        root.quit()
-        root.destroy()
-
-    #
-    def minimizer(e):
-        root.overrideredirect(0)
-        root.iconify()
+    root.minimized = False  # only to know if root is minimized
+    root.maximized = False  # only to know if root is maximized
 
     # Create New Title Bar
     title_bar = Frame(root, bg="#1F262A", relief="raised", bd=1)
     title_bar.pack(fill=X)
 
     # 'Binding the title bar
-    title_bar.bind("<B1-Motion>", move_app)
+    title_bar.bind("<Map>", frame_mapped)
 
     # Navigation Buttons
     home_button = TkinterCustomButton(master=title_bar, bg_color=None,
+                                      fg_color="#1F262A",
+                                      hover_color="#2a3439",
+                                      text_font="Bold, 14",
+                                      text="Home",
+                                      text_color="white",
+                                      corner_radius=0,
+                                      width=75,
+                                      height=40,
+                                      hover=True,
+                                      command=lambda: call_main())
+    home_button.pack(side=LEFT, padx=5)
+
+    # Create Login Button
+    login_button = TkinterCustomButton(master=title_bar, bg_color=None,
+                                       fg_color="#1F262A",
+                                       hover_color="#2a3439",
+                                       text_font="Bold, 14",
+                                       text="Login",
+                                       text_color="white",
+                                       corner_radius=0,
+                                       width=75,
+                                       height=40,
+                                       hover=True,
+                                       command=lambda: LoginPage())
+    login_button.pack(side=LEFT, padx=5)
+
+    # Results Button here
+    results_button = TkinterCustomButton(master=title_bar, bg_color=None,
+                                         fg_color="#1F262A",
+                                         hover_color="#2a3439",
+                                         text_font="Bold, 14",
+                                         text="Results",
+                                         text_color="white",
+                                         corner_radius=0,
+                                         width=75,
+                                         height=40,
+                                         hover=True,
+                                         command=lambda: ResultsPage())
+    results_button.pack(side=LEFT, padx=5)
+
+    # Create Settings Button
+    settings_button = TkinterCustomButton(master=title_bar, bg_color=None,
                                           fg_color="#1F262A",
                                           hover_color="#2a3439",
                                           text_font="Bold, 14",
-                                          text="Home",
+                                          text="Settings",
                                           text_color="white",
                                           corner_radius=0,
                                           width=75,
                                           height=40,
                                           hover=True,
-                                          command=lambda: MainWindow())
-    home_button.pack(side=LEFT, padx=5)
-
-    # Create Login Button
-    login_button = TkinterCustomButton(master=title_bar, bg_color=None,
-                                      fg_color="#1F262A",
-                                      hover_color="#2a3439",
-                                      text_font="Bold, 14",
-                                      text="Login",
-                                      text_color="white",
-                                      corner_radius=0,
-                                      width=75,
-                                      height=40,
-                                      hover=True,
-                                      command=lambda: LoginPage())
-    login_button.pack(side=LEFT, padx=5)
-
-    # Results Button here
-    results_button = TkinterCustomButton(master=title_bar, bg_color=None,
-                                      fg_color="#1F262A",
-                                      hover_color="#2a3439",
-                                      text_font="Bold, 14",
-                                      text="Results",
-                                      text_color="white",
-                                      corner_radius=0,
-                                      width=75,
-                                      height=40,
-                                      hover=True,
-                                      command=lambda: ResultsPage())
-    results_button.pack(side=LEFT, padx=5)
-
-    # Create Settings Button
-    settings_button = TkinterCustomButton(master=title_bar, bg_color=None,
-                                      fg_color="#1F262A",
-                                      hover_color="#2a3439",
-                                      text_font="Bold, 14",
-                                      text="Settings",
-                                      text_color="white",
-                                      corner_radius=0,
-                                      width=75,
-                                      height=40,
-                                      hover=True,
-                                      command=lambda: SettingsPage())
+                                          command=lambda: SettingsPage())
     settings_button.pack(side=LEFT, padx=5)
 
     # Create Help Button
@@ -446,23 +111,114 @@ def create_window():
                                       command=lambda: ResultsPage())
     scan_button.place(relx=.05, rely=.1)
 
-    # Create Title Text
-    title_label = Label(title_bar, text="Software Inventory Tool", font="Bold, 10", bg="#1F262A", fg="white")
-    title_label.place(relx=.5, rely=.5, anchor="center")
+    # Create Express Scan Button
+    express_scan_button = TkinterCustomButton(master=root, bg_color="#2a3439",
+                                      fg_color="#1F262A",
+                                      hover_color="#AAA9AD",
+                                      text_font="Bold, 14",
+                                      text="Express Scan",
+                                      text_color="white",
+                                      corner_radius=10,
+                                      width=130,
+                                      height=40,
+                                      hover=True,
+                                      command=lambda: ResultsPage())
+    express_scan_button.place(relx=.15, rely=.1)
 
-    # Create close button
-    close_label = Label(title_bar, text="X", bg="#1f262A", fg="white", font=("", 16), relief="raised", bd=0)
-    close_label.pack(side=RIGHT, padx=4, pady=4)
-    close_label.bind("<Button-1>", quitter)
+    close_button = Button(title_bar, text='  ×  ', command=root.destroy, bg="#1f262A", padx=2, pady=2,
+                          font=("calibre", 13),
+                          bd=0, fg='white', highlightthickness=0)
+    expand_button = Button(title_bar, text=' 🗖 ', bg="#1f262A", padx=2, pady=2, bd=0, fg='white',
+                           font=("calibre", 13), highlightthickness=0)
+    minimize_button = Button(title_bar, text=' — ', bg="#1f262A", padx=2, pady=2, bd=0, fg='white',
+                             font=("calibre", 13), highlightthickness=0)
+    title_bar_title = Label(title_bar, text="Software Inventory Tool", bg="#1f262A", bd=0, fg='white',
+                            font=("helvetica", 10),
+                            highlightthickness=0)
+    minimize_button.bind("<Button-1>", minimizer)
+    expand_button.bind("<Button-1>", maximize_me)
 
-    # Create Minimize button
-    minimize_label = Label(title_bar, text="─", bg="#1f262A", fg="white", font=("", 16), relief="raised", bd=0)
-    minimize_label.pack(side=RIGHT, pady=4)
-    minimize_label.bind("<Button-1>", minimizer)
+    # Packing the title_bar with all the buttons
+    title_bar.pack(fill=X)
+    close_button.pack(side=RIGHT, ipadx=7, ipady=1)
+    expand_button.pack(side=RIGHT, ipadx=7, ipady=1)
+    minimize_button.pack(side=RIGHT, ipadx=7, ipady=1)
+    title_bar_title.pack(side=LEFT, padx=10)
+
+    # Functions to change the color of buttons when hovered over
+    def changex_on_hovering(event):
+        close_button.configure(bg="red")
+
+    def returnx_to_normalstate(event):
+        close_button.configure(bg="#1f262A")
+
+    def change_size_on_hovering(event):
+        expand_button.configure(bg="#2a3439")
+
+    def return_size_on_hovering(event):
+        expand_button.configure(bg="#1f262A")
+
+    def changem_size_on_hovering(event):
+        minimize_button.configure(bg="#2a3439")
+
+    def returnm_size_on_hovering(event):
+        minimize_button.configure(bg="#1f262A")
+
+    def change_text_on_click(e):
+        if expand_button.cget("text") == " 🗗 ":
+            expand_button.configure(text=" 🗖 ")
+        else:
+            expand_button.configure(text=" 🗗 ")
+        maximize_me(e)
+
+    def get_pos(e):  # this is executed when the title bar is clicked to move the window
+
+        if not root.maximized:
+
+            xwin = root.winfo_x()
+            ywin = root.winfo_y()
+            startx = e.x_root
+            starty = e.y_root
+
+            ywin = ywin - starty
+            xwin = xwin - startx
+
+            def move_window(e):  # runs when window is dragged
+                root.config(cursor="fleur")
+                root.geometry(f'+{e.x_root + xwin}+{e.y_root + ywin}')
+
+            def release_window(e):  # runs when window is released
+                root.config(cursor="arrow")
+
+            title_bar.bind('<B1-Motion>', move_window)
+            title_bar.bind('<ButtonRelease-1>', release_window)
+            title_bar_title.bind('<B1-Motion>', move_window)
+            title_bar_title.bind('<ButtonRelease-1>', release_window)
+
+        else:
+            expand_button.config(text=" 🗖 ")
+            root.maximized = not root.maximized
+
+    title_bar.bind('<Button-1>', get_pos)  # so you can drag the window from the title bar
+    title_bar_title.bind('<Button-1>', get_pos)  # so you can drag the window from the title
+
+    # Binding buttons to Function to change color if hovered over
+    close_button.bind('<Enter>', changex_on_hovering)
+    close_button.bind('<Leave>', returnx_to_normalstate)
+    expand_button.bind('<Enter>', change_size_on_hovering)
+    expand_button.bind('<Leave>', return_size_on_hovering)
+    expand_button.bind('<Button-1>', change_text_on_click)
+    minimize_button.bind('<Enter>', changem_size_on_hovering)
+    minimize_button.bind('<Leave>', returnm_size_on_hovering)
 
     root.mainloop()
 
 
-if __name__ == '__main__':
+def call_main():
+    MainWindow()
     create_window()
-    #main = mainloop()
+
+
+if __name__ == '__main__':
+    MainWindow()
+    create_window()
